more dehydra. disabled lockfree module (not in use).
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6240 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/EntityHandles.cpp
@@ -12,9 +12,9 @@ CHandle::CHandle()
 	m_refcount = 0;
 }
 
-HEntity::HEntity( u16 index )
+HEntity::HEntity( size_t index )
 {
-	m_handle = index;
+	m_handle = (u16)index;
 	AddRef();
 }
